Made separate app for usersettings. Could not move user model.
