+++ Skidom/Skidom/urls.py
@@ -20,6 +20,7 @@ from django.contrib import admin
 urlpatterns = [
     url(r'^polls/', include('polls.urls')),
     url(r'^resorthub/', include('resorthub.urls')),
+    url(r'^usersettings/', include('usersettings.urls')),
     url(r'^admin/', admin.site.urls), 
     url(r'^accounts/', include('django.contrib.auth.urls')),
 ]
