+++ Skidom/usersettings/models.py
@@ -1,0 +1,3 @@
+from __future__ import unicode_literals
+
+from django.db import models
